+++ src/deluge/modulation/params/param_collection.cpp
@@ -39,7 +39,7 @@ void ParamCollection::notifyParamModifiedInSomeWay(ModelStackWithAutoParam const
 
 	bool currentValueChanged = (oldValue != modelStack->autoParam->getCurrentValue());
 	if (currentValueChanged || automationChanged) {
-		modelStack->paramManager->notifyParamModifiedInSomeWay(modelStack, currentValueChanged, automationChanged,
+		modelStack->paramManager->notifyParamModifiedInSomeWay(modelStack, static_cast<int32_t>(currentValueChanged), automationChanged,
 		                                                       automatedNow);
 	}
